Josephus problem
The Josephus problem is the following game: N people, numbered 1 to N,
are sitting in a circle. Starting at person 1, a hot potato is passed.
After M passes, the person holding the hot potato is eliminated, the
circle closes ranks, and the game con- tinues with the person who was
sitting after the eliminated person picking up the hot potato. The last
remaining person wins. Thus, if M = 0 and N = 5, players are eliminated
in order, and player 5 wins. If M = 1 and N = 5, the order of
elimination is 2, 4, 1, 5